Added new buildAll script to generate arguments

+++ build.sh
@@ -5,14 +5,11 @@
 #  By default this will not build the versioned images (non-latest versions), but this can be enabled by using the --buildVersionedImages.
 set -Eeo pipefail
 
-usage="Usage (all args optional): buildAll.sh --version=<version> --buildLabel=<build label> --communityRepository=<communityRepository> --officialRepository=<officialRepository>  --fullDownloadUrl=<fulldownload image download url>"
+readonly usage="Usage: build.sh --version=<version> --buildLabel=<build label> --communityRepository=<communityRepository> --officialRepository=<officialRepository>  --fullDownloadUrl=<fulldownload image download url>"
 
-version=19.0.0.11
-buildLabel=cl191120191031-0300
 communityRepository=openliberty/open-liberty
 officialRepository=open-liberty
 
-fullDownloadUrl="https://repo1.maven.org/maven2/io/openliberty/openliberty-kernel/${version}/openliberty-kernel-${version}.zip"
 
 readonly IMAGE_ROOT="releases" # the name of the dir holding all versions
 readonly REPO="openliberty/daily"
@@ -45,6 +42,12 @@ main () {
     shift
     done
 
+    if [[ -z "${buildLabel}" || -z "${fullDownloadUrl}" || -z "${version}" ]]; then
+      echo "Error: buildLabel, fullDownloadUrl, and version are required flags"
+      echo "${usage}"
+      exit 1
+    fi
+
     wget --progress=bar:force $fullDownloadUrl -U UA-Open-Liberty-Docker -O full.zip
     fullDownloadSha=$(sha1sum full.zip | awk '{print $1;}')
     rm -f full.zip
@@ -56,7 +59,7 @@ main () {
         echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_ID" --password-stdin
     fi
 
-    local tags=(kernel full)
+    local tags=(kernel)
 
     for tag in "${tags[@]}"; do
       build_latest_tag $tag
@@ -70,15 +73,15 @@ build_latest_tag() {
     # set image information arrays
     local file_exts_ubi=(adoptopenjdk8 adoptopenjdk11 adoptopenjdk13 ibmjava8)
     local tag_exts_ubi=(java8-openj9-ubi java11-openj9-ubi java13-openj9-ubi java8-ibmjava-ubi)
- 
+
     for i in "${!tag_exts_ubi[@]}"; do
-        local docker_dir="${IMAGE_ROOT}/kernel/${tag}"
+        local docker_dir="${IMAGE_ROOT}/${version}/${tag}"
         local full_path="${docker_dir}/Dockerfile.ubi.${file_exts_ubi[$i]}"
         if [[ -f "${full_path}" ]]; then
-            local build_image="${REPO}:${tag}-${tag_exts_ubi[$i]}"
+            local build_image="${REPO}:full-${tag_exts_ubi[$i]}"
 
             echo "****** Building image ${build_image}..."
-            docker build --no-cache=true -t "${build_image}" -f "${full_path}" "${docker_dir}" --build-arg LIBERTY_VERSION=${version} --build-arg LIBERTY_BUILD_LABEL=${buildLabel} --build-arg LIBERTY_SHA=${fullDownloadSha} --build-arg LIBERTY_DOWNLOAD_URL=${fullDownloadUrl}
+            docker build --no-cache=true -t "${build_image}" -f "${full_path}" --build-arg LIBERTY_VERSION=${version} --build-arg LIBERTY_BUILD_LABEL=${buildLabel} --build-arg LIBERTY_SHA=${fullDownloadSha} --build-arg LIBERTY_DOWNLOAD_URL=${fullDownloadUrl} "${docker_dir}"
             handle_results $? "${build_image}"
         else
             echo "Could not find Dockerfile at path ${full_path}"
@@ -86,20 +89,19 @@ build_latest_tag() {
         fi
     done
 
-    local docker_dir="${IMAGE_ROOT}/kernel/${tag}"
+    local docker_dir="${IMAGE_ROOT}/${version}/${tag}"
     local full_path="${docker_dir}/Dockerfile.ubuntu.adoptopenjdk8"
 
     if [[ -f "${full_path}" ]]; then
-        local ubuntu_image="${REPO}:${tag}-adoptopenjdk8"
+        local ubuntu_image="${REPO}:full-adoptopenjdk8"
 
         echo "****** Building image ${ubuntu_image}..."
-        docker build --no-cache=true -t "${ubuntu_image}" -f "${full_path}" "${docker_dir}" --build-arg LIBERTY_VERSION=${version} --build-arg LIBERTY_BUILD_LABEL=${buildLabel} --build-arg LIBERTY_SHA=${fullDownloadSha} --build-arg LIBERTY_DOWNLOAD_URL=${fullDownloadUrl}
+        docker build --no-cache=true -t "${ubuntu_image}" -f "${full_path}" --build-arg LIBERTY_VERSION=${version} --build-arg LIBERTY_BUILD_LABEL=${buildLabel} --build-arg LIBERTY_SHA=${fullDownloadSha} --build-arg LIBERTY_DOWNLOAD_URL=${fullDownloadUrl} "${docker_dir}" 
         handle_results $? "${ubuntu_image}"
     else
         echo "Could not find Dockerfile at path ${full_path}"
         exit 1
     fi
-    
 }
 
 ## push the built image if build was successful and branch is master
